+++ packages/dataviews/src/field-types/datetime.tsx
@@ -5,6 +5,7 @@ import type { DataViewRenderFieldProps, SortDirection } from '../types';
 import type { FieldType } from '../types/private';
 import RenderFromElements from './utils/render-from-elements';
 import parseDateTime from './utils/parse-date-time';
+import isValidElements from './utils/is-valid-elements';
 import {
 	OPERATOR_ON,
 	OPERATOR_NOT_ON,
@@ -15,6 +16,7 @@ import {
 	OPERATOR_IN_THE_PAST,
 	OPERATOR_OVER,
 } from '../constants';
+import isValidRequired from './utils/is-valid-required';
 
 function render( { item, field }: DataViewRenderFieldProps< any > ) {
 	if ( field.elements ) {
@@ -46,10 +48,6 @@ export default {
 	render,
 	Edit: 'datetime',
 	sort,
-	isValid: {
-		elements: true,
-		custom: () => null,
-	},
 	enableSorting: true,
 	enableGlobalSearch: false,
 	defaultOperators: [
@@ -73,4 +71,8 @@ export default {
 		OPERATOR_OVER,
 	],
 	getFormat: () => ( {} ),
+	validate: {
+		required: isValidRequired,
+		elements: isValidElements,
+	},
 } satisfies FieldType< any >;