+++ packages/dataviews/src/field-types/password.tsx
@@ -4,6 +4,11 @@
 import type { DataViewRenderFieldProps } from '../types';
 import type { FieldType } from '../types/private';
 import RenderFromElements from './utils/render-from-elements';
+import isValidRequired from './utils/is-valid-required';
+import isValidMinLength from './utils/is-valid-min-length';
+import isValidMaxLength from './utils/is-valid-max-length';
+import isValidPattern from './utils/is-valid-pattern';
+import isValidElements from './utils/is-valid-elements';
 
 function render( { item, field }: DataViewRenderFieldProps< any > ) {
 	return field.hasElements ? (
@@ -18,13 +23,16 @@ export default {
 	render,
 	Edit: 'password',
 	sort: () => 0, // Passwords should not be sortable for security reasons
-	isValid: {
-		elements: true,
-		custom: () => null,
-	},
 	enableSorting: false,
 	enableGlobalSearch: false,
 	defaultOperators: [],
 	validOperators: [],
 	getFormat: () => ( {} ),
+	validate: {
+		required: isValidRequired,
+		pattern: isValidPattern,
+		minLength: isValidMinLength,
+		maxLength: isValidMaxLength,
+		elements: isValidElements,
+	},
 } satisfies FieldType< any >;